[#19] Rename: 파일명 camelCase 1차 변경
