+++ src/pages/signup/aaccountCheck.tsx
